2.5 - Separated about component and fixed related component styles
