+++ src/app/components/about/about.component.html
@@ -1,0 +1,8 @@
+<div class="app-head">
+    <h2>About</h2>
+</div>
+<div class="app-content">
+    <p>This is a simple youtube web player and desktop player based on angular-cli, socket.io, ngu-carousel, @angular/material.</p>
+    <p>This player is compatible only with Chrome/webkit browsers.</p>
+    <p>The project status is beta.</p>
+</div>
