+++ src/app/components/about/about.component.scss
@@ -1,0 +1,10 @@
+app-about {
+    .app-content {
+        background-color: rgba(0, 0, 0, 0.3);
+        padding: 10px 20px;
+        border-radius: 5px;
+        p {
+            margin: 5px 0;
+        }
+    }
+}
