+++ src/app/components/about/about.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit, ViewEncapsulation } from '@angular/core';
+
+@Component({
+  selector: 'app-about',
+  templateUrl: 'about.component.html',
+  styleUrls: ['./about.component.scss'],
+  encapsulation: ViewEncapsulation.None
+})
+
+export class AboutComponent implements OnInit {
+
+  ngOnInit() {
+    console.log('about');
+  }
+
+}
