+++ src/app/components/player/video-item/video-item.component.scss
@@ -1,0 +1,137 @@
+.video-item {
+    display: flex;
+    align-items: center;
+    background-color: var(--c-dark-gray);
+    color: var(--c-white);
+    padding: 0;
+    position: relative;
+    transition: all 0.1s ease-in-out;
+    h2 {
+        font-size: 14px;
+        margin: 0;
+    }
+    &-autor {
+        font-size: 12px;
+        color: var(--c-white);
+    }
+    &-image {
+        display: block;
+        height: 70px;
+        width: 80px;
+        background-position: center center;
+        margin-right: 10px;
+        transition: all 0.1s ease-out;
+        border-radius: 6px;
+    }
+    .video-item-details {
+        display: flex;
+        color: rgba(255, 255, 255, 0.8);
+        font-weight: 400;
+        font-size: 12px;
+        p {
+            margin: 0 15px 0 0;
+        }
+        .fa {
+            margin-right: 3px;
+        }
+    }
+    &-content {
+        width: 100%;
+        overflow: hidden;
+        position: relative;
+        p {
+            margin: 0;
+            white-space: nowrap;
+            text-overflow: ellipsis;
+            overflow: hidden;
+        }
+    }
+    &-info {
+        display: flex;
+        align-items: center;
+        position: relative;
+        z-index: 1;
+        width: 100%;
+    }
+    &-settings {
+        color: var(--c-white);
+        position: absolute;
+        right: 0;
+        top: 0;
+        height: 100%;
+        width: 100%;
+        z-index: 1;
+        opacity: 0;
+        visibility: hidden;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        transition: all 0.1s ease-in-out;      
+        > button {
+            background-color: #252729;
+            color: inherit;
+            height: 30px;
+            width: 30px;
+            border-radius: 50%;
+            text-align: center;
+            border: none;
+            margin: 0 5px;
+            padding: 0;
+            transition: all 0.1s ease-in-out;
+            &:hover {
+                color: var(--c-black);
+                background-color: var(--c-white);
+            }
+            &:active {
+                color: var(--c-black);
+                background-color: var(--c-soft-white);
+            } 
+        }
+        .video-item-play {
+            .fa {
+                font-size: 12px;
+            }
+            .fa-play {
+                margin-left: 2px;
+            }
+            .fa-pause {
+                margin-left: 1px;
+            }
+        }
+        .video-item-add {
+            .fa {
+                position: relative;
+            }
+        }
+    }
+    &-title {
+        line-height: 18px;
+    }
+    &-hint {
+        text-transform: uppercase;
+        font-size: 10px;
+        line-height: 16px;
+        border: 1px solid var(--c-primary);
+        color: var(--c-primary);
+        display: none;
+        padding: 0 5px;
+        font-weight: 600 !important;
+        .fa {
+            font-size: 8px;
+            margin-right: 3px;
+            position: relative;
+            top: -1px;
+        }
+    }
+    &:hover {
+        .video-item-settings {
+            opacity: 1;
+            visibility: visible;
+        }
+    }
+    &.active {
+        .video-item-hint {
+            display: inline-block;
+        }
+    }
+}
