+++ src/app/components/related/related.component.html
@@ -16,16 +16,7 @@
     <ngu-carousel
       [inputs]="carouselOne" [dataSource]="globals.relatedVideos">
       <ngu-item *nguCarouselDef="let relatedVideo; let i = index;">
-        <div [attr.data-index]="i" class="video-item">
-          <div class="video-item-info">
-            <div *ngIf="globals.thumbnails" class="video-item-image" [ngStyle]="{'background-image': 'url(' + relatedVideo.thumbnails.high +')'}"></div>
-            <div class="video-item-content">
-              <p class="video-item-title">{{relatedVideo.title}}</p>
-              <p class="video-item-autor">{{relatedVideo.channelTitle}}</p>
-            </div>
-            <app-buttons [videoIndex]="i" [listID]="2"></app-buttons>
-          </div>
-        </div>
+        <app-video-item [videoIndex]="i" [videoItem]="relatedVideo" [listID]="2"></app-video-item>
       </ngu-item>
       <button NguCarouselPrev class='leftRs'><i class="fa fa-angle-left"></i></button>
       <button NguCarouselNext class='rightRs'><i class="fa fa-angle-right"></i></button>
