--- src/app/components/youtube-about.component.ts
@@ -1,14 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-about',
-  templateUrl: 'youtube-about.component.html'
-})
-
-export class AboutComponent implements OnInit {
-
-  ngOnInit() {
-    console.log('about');
-  }
-
-}
