+++ src/assets/scss/_common.scss
@@ -164,41 +164,3 @@ h2 {
         border-radius: 30px;
     }
 }
-
-.session-section {
-    margin-bottom: 20px;
-    input[type="text"] {
-        background: transparent;
-        border: 1px solid #e4e4e4;
-        border-radius: 3px;
-        padding: 8px 10px;
-        font-size: $f-14;
-    }
-    .session-key-block {
-        display: block;
-        margin: 0 auto 10px;
-        label {
-            margin-right: 5px;
-        }
-        input {
-            width: 100%;
-        }
-    }
-}
-
-app-room {
-    display: inline-block;
-    width: 100%;
-    margin-bottom: 15px;
-}
-
-app-about {
-    .app-content {
-        background-color: rgba(0, 0, 0, 0.3);
-        padding: 10px 20px;
-        border-radius: 5px;
-        p {
-            margin: 5px 0;
-        }
-    }
-}
