+++ src/assets/scss/_videos.scss
@@ -1,28 +1,3 @@
-.video-list {
-    width: 100%;
-    overflow-y: auto;
-    &-info {
-        text-align: center;
-        margin: 5px;
-    }
-    &::-webkit-scrollbar-track {
-        border-radius: 2px;
-    }
-
-    &::-webkit-scrollbar {
-        width: 6px;
-        background-color: var(--c-dark-gray);
-    }
-
-    &::-webkit-scrollbar-thumb {
-        border-radius: 20px;
-        background-color: rgba(#FFF, 0.1);
-        &:hover {
-            background-color: rgba(#FFF, 0.4);
-        }
-    }
-}
-
 .video-item-head {
     position: relative;
     margin: 0;
@@ -54,144 +29,6 @@
     }
     .fa {
         margin-right: 3px;
-    }
-}
-
-.video-item {
-    display: flex;
-    align-items: center;
-    background-color: var(--c-dark-gray);
-    color: var(--c-white);
-    padding: 0;
-    position: relative;
-    transition: all 0.1s ease-in-out;
-    h2 {
-        font-size: $f-14;
-        margin: 0;
-    }
-    &-autor {
-        font-size: $f-12;
-        color: var(--c-white);
-    }
-    &-image {
-        display: block;
-        height: 70px;
-        width: 80px;
-        background-position: center center;
-        margin-right: 10px;
-        transition: all 0.1s ease-out;
-        border-radius: 6px;
-    }
-    .video-item-details {
-        display: flex;
-        color: rgba(255, 255, 255, 0.8);
-        font-weight: 400;
-        font-size: $f-12;
-        p {
-            margin: 0 15px 0 0;
-        }
-        .fa {
-            margin-right: 3px;
-        }
-    }
-    &-content {
-        width: 100%;
-        overflow: hidden;
-        position: relative;
-        p {
-            margin: 0;
-            white-space: nowrap;
-            text-overflow: ellipsis;
-            overflow: hidden;
-        }
-    }
-    &-info {
-        display: flex;
-        align-items: center;
-        position: relative;
-        z-index: 1;
-        width: 100%;
-    }
-    &-settings {
-        color: var(--c-white);
-        position: absolute;
-        right: 0;
-        top: 0;
-        height: 100%;
-        width: 100%;
-        z-index: 1;
-        opacity: 0;
-        visibility: hidden;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        transition: all 0.1s ease-in-out;      
-        > button {
-            background-color: #252729;
-            color: inherit;
-            height: 30px;
-            width: 30px;
-            border-radius: 50%;
-            text-align: center;
-            border: none;
-            margin: 0 5px;
-            padding: 0;
-            transition: all 0.1s ease-in-out;
-            &:hover {
-                color: var(--c-black);
-                background-color: var(--c-white);
-            }
-            &:active {
-                color: var(--c-black);
-                background-color: var(--c-soft-white);
-            } 
-        }
-        .video-item-play {
-            .fa {
-                font-size: $f-12;
-            }
-            .fa-play {
-                margin-left: 2px;
-            }
-            .fa-pause {
-                margin-left: 1px;
-            }
-        }
-        .video-item-add {
-            .fa {
-                position: relative;
-            }
-        }
-    }
-    &-title {
-        line-height: 18px;
-    }
-    &-hint {
-        text-transform: uppercase;
-        font-size: $f-10;
-        line-height: 16px;
-        border: 1px solid var(--c-primary);
-        color: var(--c-primary);
-        display: none;
-        padding: 0 5px;
-        font-weight: 600 !important;
-        .fa {
-            font-size: $f-8;
-            margin-right: 3px;
-            position: relative;
-            top: -1px;
-        }
-    }
-    &:hover {
-        .video-item-settings {
-            opacity: 1;
-            visibility: visible;
-        }
-    }
-    &.active {
-        .video-item-hint {
-            display: inline-block;
-        }
     }
 }
